Add 'ImplicitSkyMapArrayRef' and start implementing mom merge method.

+++ src/nested/map/fits/write.rs
@@ -22,7 +22,7 @@ pub fn write_implicit_skymap_fits<R: Write, T: SkyMapValue>(
   // since we already know it.
   // In the end, using monomorphisation, the binary code is duplicated so...
   let n_cells = values.len() as u64;
-  let depth = depth_from_n_hash_unsafe(values.len());
+  let depth = depth_from_n_hash_unsafe(values.len() as u64);
   if n_cells != n_hash(depth) {
     return Err(FitsError::new_custom(format!(
       "Number of cell {} not compatible with an HEALPix depth of {}. Expected: {}.",
